requirements.txt
fastmcp>=0.1.0
gitpython>=3.1.0
requests>=2.25.0